pyproject.toml
[tool.poetry]
name = "memgraph"
version = "0.1.0"
description = ""
authors = ["geforcemcdoodle <h0091891@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cmake = "^3.30.3"
gqlalchemy = "^1.6.0"
fastapi = "^0.115.0"
strawberry-graphql = "^0.242.0"
pytest = "^8.3.3"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
